Never use negative-sized scissor.
As the size of a scissor rectangle used in ScrollArea is defined as (width - something, height - something),
the rectangle would have negative size if e.g. the scissor area size is proportional to window size and the
window gets resized to a small size. This would result in a segfault (at least on AMD Catalyst/Linux).

E.g. when a user plays around with window width/height the program (even dimgui examples) would crash.

I worked around this by forcing the scissor area to be at least 1x1 pixels large.